backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
loguru==0.7.2
httpx==0.27.2
pytest==8.3.4
asyncpg==0.30.0
greenlet==3.1.1
python-multipart==0.0.19
email-validator==2.2.0
